requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arabic-reshaper==3.0.0
arrow==1.3.0
asn1crypto==1.5.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
cssselect2==0.8.0
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exponent_server_sdk==2.2.0
fastapi==0.115.12
frozenlist==1.7.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
ics==0.7.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
lxml==6.0.0
MarkupSafe==3.0.2
multidict==6.6.4
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==6.0.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
stripe==12.4.0
svglib==1.5.1
TatSu==5.13.1
tinycss2==1.4.0
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.34.3
webencodings==0.5.1
wrapt==1.17.2
xhtml2pdf==0.2.17
yarl==1.20.1
icalendar==6.3.1
gunicorn
redis
Werkzeug==3.1.3
